requirements.txt
aiogram==3.7.0
opencv-python==4.9.0.80
google-generativeai==0.8.3
ezdxf==1.3.4
svgwrite==1.4.3
aiosqlite==0.20.0
python-dotenv==1.0.1
Pillow==10.3.0
numpy==1.26.4
